+++ client/components/tasks/TaskLabelsInCard.vue
@@ -1,7 +1,9 @@
 <template>
 
 <div class="task-labels">
-  <div class="label" v-for="label in labels" :key="label._id" :style="getColor(label)"></div>
+  <div :title="label.name" :class="{ label: true, show: showLabelText }" v-for="label in labels" :key="label._id" :style="getColor(label)" @click.stop="toggleLabel">
+    <template v-if="showLabelText">{{ label.name }}</template>
+  </div>
 </div>
 
 </template>
@@ -9,6 +11,7 @@
 <script>
 import { Labels } from '/imports/api/labels/labels.js'
 import { Projects } from '/imports/api/projects/projects.js'
+import { mapState } from "vuex";
 
 export default {
   props: {
@@ -16,6 +19,11 @@ export default {
       type: Object,
       default: {}
     }
+  },
+  computed: {
+    ...mapState([
+      "showLabelText",
+    ]),
   },
   data() {
     return {
@@ -38,6 +46,10 @@ export default {
   methods: {
     getColor (label) {
       return 'background-color: ' + label.color;
+    },
+
+    toggleLabel (e) {
+      this.$store.dispatch("setShowLabelText", !this.showLabelText)
     }
   }
 };
@@ -54,5 +66,17 @@ export default {
   width: 48px;
   height: 8px;
   margin-right: 4px;
+}
+
+.show {
+  width: auto;
+  min-width: 48px;
+  max-width: 100px;
+  height: 16px;
+  margin-right: 4px;
+  font-size: 9px;
+  padding: 2px;
+  overflow-x: hidden;
+  text-align: center;
 }
 </style>