+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Label text can be displayed by clicking on it
 
 ### Changed
 - Fix task filter on dashboard when not admin